refactoring of Object and Const
