+++ src/app/Components/amy-chat/amy-chat.component.ts
@@ -3,8 +3,8 @@ import { DatabaseService } from '../../Services/database.service';
 import { TTSService } from '../../Services/tts.service';
 import { SpeechRecognitionService } from '../../Services/speechrecognition.service';
 import { Command } from '../../Objects/command';
-import { AMY_UNKNOWN_COMMAND_RESPONSE, CHAT_DISPLAY_BUTTON_ACTIVE, CHAT_DISPLAY_BUTTON_INACTIVE, COMMAND_INPUT_PLACEHOLDER, USER_CHAT_NAME, AMY_CHAT_NAME} from './strings';
-import { Message } from "./message";
+import { AMY_UNKNOWN_COMMAND_RESPONSE, CHAT_DISPLAY_BUTTON_ACTIVE, CHAT_DISPLAY_BUTTON_INACTIVE, COMMAND_INPUT_PLACEHOLDER, USER_CHAT_NAME, AMY_CHAT_NAME} from '../../Constants/strings';
+import { Message } from "../../Objects/message";
 import {animate, state, style, transition, trigger} from '@angular/animations';
 import { interval } from 'rxjs';
 import {mergeMap} from 'rxjs/operators';
